Move fhir-ig utils out of src/main/java and into a '.tool' package
