+++ conformance/fhir-ig-davinci-pdex-formulary/src/test/java/com/ibm/fhir/ig/davinci/pdex/formulary/tool/ResourceProcessor.java
@@ -1,10 +1,10 @@
 /*
- * (C) Copyright IBM Corp. 2020
+ * (C) Copyright IBM Corp. 2020, 2021
  *
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package com.ibm.fhir.ig.davinci.pdex.formulary.util;
+package com.ibm.fhir.ig.davinci.pdex.formulary.tool;
 
 import java.io.BufferedReader;
 import java.io.File;
