+++ conformance/fhir-ig-davinci-pdex-plan-net/src/test/java/com/ibm/fhir/ig/davinci/plannet/tool/ResourceProcessor.java
@@ -1,10 +1,9 @@
 /*
- * (C) Copyright IBM Corp. 2020
+ * (C) Copyright IBM Corp. 2020, 2021
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-
-package com.ibm.fhir.ig.davinci.pdex.plannet.util;
+package com.ibm.fhir.ig.davinci.plannet.tool;
 
 import java.io.BufferedReader;
 import java.io.File;
