+++ conformance/fhir-ig-mcode/src/test/java/com/ibm/fhir/ig/mcode/tool/IndexGenerator.java
@@ -1,10 +1,10 @@
 /*
- * (C) Copyright IBM Corp. 2020
+ * (C) Copyright IBM Corp. 2020, 2021
  *
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package com.ibm.fhir.ig.mcode.util;
+package com.ibm.fhir.ig.mcode.tool;
 
 import java.io.File;
 import java.io.FileOutputStream;
